requirements.txt
fastapi
uvicorn
mysql.connector
streamlit
nmap